Suppress warning of lemon.c

+++ src/currentia/query/tools/lemon/lemon.c
@@ -2523,7 +2523,7 @@ void Parse(struct lemon *gp)
     fclose(fp);
     return;
   }
-  if( fread(filebuf,1,filesize,fp)!=filesize ){
+  if( fread(filebuf,1,filesize,fp)!=(size_t)filesize ){
     ErrorMsg(ps.filename,0,"Can't read in all %d bytes of this file.",
       filesize);
     free(filebuf);
